Move dnsmasq installation after Microshift installation
According to the relase notes [1] for Microshift 4.13, normally the
it should be fine, when the dnsmasq service will be started before
Microshift installation. When I do tests, it seems that for Microshift
4.12 it is better, when the service is deployed after the Microshift
installation - there is no error in openshift-dns pod:

    oc logs -n openshift-dns -l dns.operator.openshift.io/daemonset-dns=default

We would like to support both releases and this change should also work on
newest Microshift deployment.

[1] https://github.com/openshift/openshift-docs/pull/58946/files