Listen dnsmasq also on local ip address
This change will enable on dnsmasq listening in NetworkManager also on the
on local ip address of the host. It would be useful, when
someone would like to provide local domains, that can be resolved by the
dnsmasq service inside the OpenShift cluster.
For example, if someone would like to enable resolving hostname:
logserver.sftests.com inside the pod, not by using the Kubernetes like
domain name resolution, it is needed to add a new entry into the dns-default
configmap and the dns query will be forwarded to the dnsmasq on the
host.
Example:

    oc edit configmap -n openshift-dns  dns-default

    (...)
       logserver.sftest.com {
         errors
         cache 30
         forward . <server local ip address that is in dnsmasq.conf>
       }

Then  restart services

    oc -n openshift-dns rollout restart daemonsets dns-default
    oc -n openshift-dns rollout restart daemonsets node-resolver