Add support to forward external domains DNS queries to dnsmasq
That feature will be mostly used by the CI for simulating the external
domains that exists normaly in the Internet.
For example, if domain '.sftests.com' have many subdomains and not
each domain is registered domain in the Openshift cluster and the domain
is just for test (not registered in any DNS authority) it will fail
name resolution inside the pod with an error:

    "curl: (6) Could not resolve host: test.sftests.com"

In that case, we are adding a forward option in the configuration configmap,
which will forward the DNS queries made for that domain to the dnsmasq,
which is installed on the host. The dnsmasq (which is started by NetworkManager)
have proper entries to resolve the domain.