Set lvm volumes to be active after reboot
After the reboot, the volume group is available:

    $ vgdisplay
      --- Volume group ---
      VG Name               rhel
      (...)

and the volumes are also available:

    $ lvdisplay
      --- Logical volume ---
      LV Path                /dev/rhel/37449bb2-9190-4846-8ec1-562361e0fdbc
      LV Name                37449bb2-9190-4846-8ec1-562361e0fdbc
      (...)

but the directory /dev/rhel is not available, so by checking:

    $ lvscan
      inactive          '/dev/rhel/37449bb2-9190-4846-8ec1-562361e0fdbc' [1,00 GiB] inherit
      inactive          '/dev/rhel/48917243-f3e8-46b5-be4d-8b40b5e0a8ec' [1,00 GiB] inherit
      (...)

shows that all volumes are inactive state.
By iterating on all inactive volumes and execute command: vchange -ay <volume>
starts the volume mapping and the /dev/rhel is available. After a while,
all pods that uses pv are starting.