Drop changing coredns configmap
After talk with a Microshift team, the core Openshift services should
not be modified. Changing the DNS to custom can be done by
systemd-resolved, which is automatically detected by the Microshift
service [1].
On the Centos 9 Stream, systemd-resolved is not installed and configured
by default, but we already configure the NetworkManager to use the
dnsmasq. In that case, let's point in the /etc/resolv.conf file, that
the nameserver is a host ip address and mark that file that should not
be changed by the NetworkManager.

[1] https://github.com/openshift/microshift/blob/df6387ed015a14018d6ab412bfea34b326a6feb4/pkg/components/controllers.go#L234-L235