Change repository directory
This commit is more a cosmetic change, because packages for few releases
are same in both location - ocp and ocp-dev-preview, but it might
happen, that there would be different version there.
To avoid situation, that we would be able to install preview version,
let's change the repository url to official.