Add lvm information for storageclass
It seems that in the new MicroShift version, the volume group (vg) name
has changed and the default value is not "rhel" as it was.
Even when we set the value to "rhel" to avoid situation, that on some
hosts when we perform an upgrade we can loose data, the MicroShift is
also creating vg with name "sc".