Enable fast-datapath-for-rhel subscription in old way
On RHEL system, if someone wants to install the MicroShift in old way,
it will miss required package: centos-release-nfv-openvswitch
that is necessary to deploy MicroShift (it provides Open vSwitch).
The package installation requires few other packages, that might make
the RHEL system unstable, so it is more safe to enable fast-datapath-for-rhel
subscription.

This is a temporary workaround until we drop CentOS support for
that project (only RHOCP version would be available).